+++ TetrisAPI/EBlockType.h
@@ -19,7 +19,7 @@ namespace TetrisAPI
 	{
 		std::vector<EBlockType> blocks = { I,L,J,Z,S,T,O };
 		srand(static_cast<unsigned int>(time(nullptr)));
-		int randomIndex = rand() % (sizeof(blocks) / sizeof(blocks[0]));
+		uint16_t randomIndex = rand() % (sizeof(blocks) / sizeof(blocks[0]));
 
 		return blocks[randomIndex];
 	}
